Use known bounds to improve division
The numbers we deal with are at every point known to lie in a range
such as (0, 10^8), (0,10^6), (0,10^4), (0,10^2). For the former two,
we can implement division by 100 by doing a 32x32->64 bit
multiplication by 0x28f5c29 followed by a right-shift by 32 (which,
depending on the cpu, may be free, if the result of the multiplication
is stored in a pair of 32 bit registers). In the third case, 32x32->32
multiplication by 0x147b followed by right-shift of 19 is sufficient.

Compared to the initial code where gcc did the /100 and %100, we win
somewhat more when the numbers are large (uniformly distributed and
also in the first few neg_binom cases). Unsurprisingly, the gain is
almost the same when the numbers are mostly 4-7 bits in size, since in
that case we follow the same fastpath and end up with a single table
lookup and no multiplications/divisions at all.

Distribution              Function             Cycles       Conversions/1 sec
uniform([10, 2^64-1])     linux_put_dec           264605076             23036508
uniform([10, 2^64-1])     rv_put_dec              123184548             45938121
                          +/-                       -53.45%              +99.41%
3 + neg_binom(0.05)       linux_put_dec           127590580             45720564
3 + neg_binom(0.05)       rv_put_dec               86442001             64071490
                          +/-                       -32.25%              +40.14%
3 + neg_binom(0.10)       linux_put_dec            93851448             61493386
3 + neg_binom(0.10)       rv_put_dec               66515632             81424562
                          +/-                       -29.13%              +32.41%
3 + neg_binom(0.15)       linux_put_dec            76263973             74853033
3 + neg_binom(0.15)       rv_put_dec               56078830             95161980
                          +/-                       -26.47%              +27.13%
3 + neg_binom(0.20)       linux_put_dec            65207687             87112918
3 + neg_binom(0.20)       rv_put_dec               48791752            109202155
                          +/-                       -25.17%              +25.36%
3 + neg_binom(0.50)       linux_put_dec            36725428            159632668
3 + neg_binom(0.50)       rv_put_dec               27447096            203274148
                          +/-                       -25.26%              +27.34%