test: prepare for timing on non-x86
rdtsc() is an x86ism; I want to try to make this easy to compile and
run on other platforms as well. So abstract the type used to record an
absolute point in time, and use signed long long for expressing a
delta-time in some platform-defined unit (e.g. nanoseconds if we end
up using clock_gettime).