add dependency generation
Generate a few .deps files so that I don't have to type make -B all
the time. We don't want the .h files to show up on the command line
for compiling the final binary, so filter them out. I'm sure there's a
better way than this but life is too short to figure that out.