remove GSL dependency
We don't use much from the GNU Scientific Library. The sort-of
negative binomial distribution we use is easy enough to roll by hand,
more or less just based on its very definition ("number of failures
before a success in series of independent trials each with probability
p of success"). GSL internally uses the Mersenne Twister generator, so
we lose some quality by switching to the *48 family, but I don't think
that matters in this case. Dropping GSL requirement makes it easier to
compile and run on other machines.