Makefile overhaul
Just build binaries called "test" and "verify". We try to auto-detect
whether to use the 32 or 64 bit code using getconf LONG_BIT - if that
fails, the user can just do

  make LONG_BIT=n

Also, don't unconditionally add -m32/-m64 to the CFLAGS, as not all
platforms support these. We still need -m32 to build the 32 bit
version on x86_64, so steal the cc-option magic from the kernel to add
-mn when it is supported.